% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{diagonal_dominance_cpp}
\alias{diagonal_dominance_cpp}
\title{Proportion of Diagonally Dominant Rows}
\usage{
diagonal_dominance_cpp(mat)
}
\arguments{
\item{mat}{Numeric matrix (assumed to be square).}
}
\value{
A double representing the proportion of rows that satisfy diagonal dominance.
}
\description{
Computes the proportion of rows in a square matrix that satisfy diagonal dominance.
}
